- Split off (renamed) SlackWeb - Use of SlackWeb for get_url - remove channel_id method
